move model builder to separate file
